+++ src/client/store/ui/assessmentFiles/index.ts
@@ -1,2 +1,6 @@
+export { useAssessmentFiles } from './hooks'
+export { useDeleteAssessmentFile } from './hooks/useDeleteAssessmentFile'
+export { useGetAssessmentFiles } from './hooks/useGetAssessmentFiles'
+export { useUploadAssessmentFile } from './hooks/useUploadAssessmentFile'
 export { AssessmentFilesActions } from './slice'
 export type { AssessmentFilesState } from './stateType'
